customize logger, handling some errors

+++ controllers/logout.js
@@ -2,10 +2,19 @@ const { client } = require('../model/Store');
 
 const logout = async (ctx) => {
   const { username } = ctx.request.body;
-  await client.del(username);
-  ctx.body = {
-    message: 'you are login off',
-  };
+  if (username === 'undefined' || username === '') {
+    ctx.status = 403;
+    return 0;
+  }
+  const candidate = await client.del(username);
+  if (candidate) {
+    ctx.body = {
+      message: 'you are login off',
+    };
+  } else {
+    ctx.message = 'user not found';
+  }
+  return 0;
 };
 
 module.exports = logout;
